+++ src/Symfony/Component/Uid/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add `UidBundle`, which provides the `uid` configuration and the services previously provided by `FrameworkBundle` under `framework.uid`
  * Add `TimeOrderedUidInterface`, implemented by `Ulid`, `UuidV6` and `UuidV7`
  * Add `Ulid::createBoundaries()`, `UuidV6::createBoundaries()` and `UuidV7::createBoundaries()` returning the lowest and highest ids sharing a timestamp